moved EPSILON, UPDATES_PER_SECOND and SUB_STEPS to Settings.kt
